Update version to 0.11.7 and refactor webview handling in VS Code extension
- Bump version in package.json and package-lock.json to 0.11.7.
- Introduce WebviewMessageHandler and WebviewContentProvider classes to streamline message handling and HTML content generation for the webview.
- Refactor CoAuthorViewProvider to utilize the new classes, improving code organization and maintainability.
- Clean up and optimize imports across multiple files for better clarity.
